pyproject.toml
[tool.poetry]
name = "mozio_transport_suppliers"
version = "0.1.0"
description = "A web app that enables transport suppliers to define custom polygons as their \"service area\" to simplify the process of capturing zip codes, cities etc"
authors = ["Njokosi Kawunju <njokosi@gmail.com>"]
license = "ISC"

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.5"
djangorestframework = "^3.13.1"
Markdown = "^3.3.7"
django-filter = "^21.1"
dj-database-url = "^0.5.0"
dj-rest-auth = "^2.2.4"
django-allauth = "^0.51.0"
django-guardian = "^2.4.0"
drf-yasg = "^1.20.0"
django-cors-headers = "^3.13.0"
django-environ = "^0.9.0"
django-axes = "^5.35.0"
djangorestframework-simplejwt = "^5.2.0"
whitenoise = "^6.2.0"

[tool.poetry.dev-dependencies]
black = "^22.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
